Splitted controller into data and backup, changed api path
